Project Files/Backend/requirements.txt
asgiref==3.8.1
blinker==1.9.0
bottle==0.13.4
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
cryptography==44.0.2
Django==5.1.5
django-cors-headers==4.9.0
django-grappelli==4.0.2
djangorestframework==3.16.1
firebase-admin==6.6.0
Flask==3.1.2
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgpack==1.1.0
packaging==25.0
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise

